+++ examples/dataplatform/product/main-datasource.tf
@@ -1,0 +1,3 @@
+data "vkcs_dataplatform_product" "spark" {
+  product_name = "spark"
+}